refactor: split file for unity migration process
